Fix testing
